requirements.txt
altgraph==0.17.5
charset-normalizer==3.4.5
customtkinter==5.2.2
darkdetect==0.8.0
et_xmlfile==2.0.0
numpy==2.4.3
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
pefile==2024.8.26
pillow==12.1.1
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.3
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
reportlab==4.4.10
setuptools==82.0.1
six==1.17.0
tzdata==2025.3
